+++ libs/shared/src/lib/components/record-history/record-history.module.ts
@@ -18,6 +18,7 @@ import { IndicatorsModule } from '@progress/kendo-angular-indicators';
 import { EmptyModule } from '../ui/empty/empty.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FormsModule } from '@angular/forms';
+import { ResizableModule } from 'angular-resizable-element';
 
 /**
  * RecordHistoryModule is a class used to manage all the modules and components
@@ -42,6 +43,7 @@ import { FormsModule } from '@angular/forms';
     FormsModule,
     TableModule,
     ToggleModule,
+    ResizableModule,
   ],
   exports: [RecordHistoryComponent],
 })