+++ libs/shared/src/lib/components/ui/map/map.component.ts
@@ -52,6 +52,7 @@ import {
   difference,
   isEqual,
   flatMapDeep,
+  pick,
 } from 'lodash';
 import {
   BehaviorSubject,
@@ -154,8 +155,6 @@ export class MapComponent
   private arcGisWebMap: any;
   /** Layer control buttons */
   private layerControlButtons: any;
-  /** Applied dashboard filters */
-  private appliedDashboardFilters: Record<string, any>;
   /** Current layer ids */
   private layerIds: string[] = [];
   /** Resize observer on map container */
@@ -206,7 +205,6 @@ export class MapComponent
     super();
     this.esriApiKey = environment.esriApiKey;
     this.mapId = uuidv4();
-    this.appliedDashboardFilters = this.contextService.filter.getValue();
   }
 
   /** Once template is ready, build the map. */
@@ -254,10 +252,15 @@ export class MapComponent
       this.contextService.filter$
         .pipe(
           debounceTime(500),
-          filter(() => {
-            const filters = this.contextService.filter.getValue();
-            return !isEqual(filters, this.appliedDashboardFilters);
-          }),
+          filter(({ previous, current }) =>
+            this.contextService.shouldRefresh(
+              this.layers.map((layer) => {
+                return pick(layer, ['datasource', 'contextFilters', 'at']);
+              }),
+              previous,
+              current
+            )
+          ),
           concatMap(() => loadNextFilters()),
           takeUntil(this.destroy$)
         )
@@ -714,19 +717,22 @@ export class MapComponent
    * Setup and draw layers on map and sets the baseTree.
    *
    * @param layerIds layerIds from saved edit layer info
+   * @param displayLayers boolean that controls whether all layers should be displayed directly
    * @returns layers
    */
-  private async getLayers(layerIds: string[]) {
+  private async getLayers(layerIds: string[], displayLayers = true) {
     /**
      * Parses a layer into a tree node
      *
      * @param layer The layer to create the tree node from
      * @param leafletLayer The leaflet layer previously created by the parent layer, if any
+     * @param displayLayers boolean that controls whether layers should be directly displayed
      * @returns The tree node
      */
     const parseTreeNode = async (
       layer: Layer,
-      leafletLayer?: L.Layer
+      leafletLayer?: L.Layer,
+      displayLayers = true
     ): Promise<OverlayLayerTree> => {
       // Add to the layers array if not already added
       if (this.layers.find((l) => l.id === layer.id)) return {} as any;
@@ -735,14 +741,14 @@ export class MapComponent
 
       if (layer.type === 'GroupLayer') {
         const children = layer.getChildren();
-        const childrenPromisse = children.map((Childrenlayer) => {
+        const childrenPromises = children.map((Childrenlayer) => {
           return this.mapLayersService
             .createLayersFromId(Childrenlayer, this.injector)
             .then(async (sublayer) => {
               if (sublayer.type === 'GroupLayer') {
                 const layer = await sublayer.getLayer();
-                return parseTreeNode(sublayer, layer);
-              } else return parseTreeNode(sublayer);
+                return parseTreeNode(sublayer, layer, displayLayers);
+              } else return parseTreeNode(sublayer, undefined, displayLayers);
             });
         });
 
@@ -752,7 +758,7 @@ export class MapComponent
           selectAllCheckbox: true,
           children:
             children.length > 0
-              ? await Promise.all(childrenPromisse)
+              ? await Promise.all(childrenPromises)
               : undefined,
         };
       } else {
@@ -763,7 +769,7 @@ export class MapComponent
         // Adds the layer to the map if not already added
         // note: group layers are of type L.LayerGroup
         // so we should check if the layer is not already added
-        if (!this.map.hasLayer(featureLayer)) {
+        if (!this.map.hasLayer(featureLayer) && displayLayers) {
           this.map.addLayer(featureLayer);
         }
         // It is a node, it does not have any children but it displays a layer
@@ -779,11 +785,11 @@ export class MapComponent
         return this.mapLayersService
           .createLayersFromId(id, this.injector)
           .then((layer) => {
-            return parseTreeNode(layer);
+            return parseTreeNode(layer, undefined, displayLayers);
           });
       });
 
-      Promise.all(layerPromises).then((layersTree) => {
+      Promise.all(layerPromises).then((layersTree: any) => {
         this.refreshLastUpdate();
         resolve({ layers: layersTree });
       });
@@ -1069,9 +1075,7 @@ export class MapComponent
   /** Set the new layers based on the filter value */
   private async filterLayers() {
     this.document.getElementById('layer-control-button-close')?.click();
-    const filters = this.contextService.filter.getValue();
     this.refreshingLayers.next(false);
-    this.appliedDashboardFilters = filters;
     const { layers: layersToGet, controls } = this.extractSettings();
 
     const shouldDisplayStatuses: Record<string, boolean> = {};
@@ -1095,7 +1099,7 @@ export class MapComponent
 
     // get new layers, with filters applied
     this.resetLayers();
-    const l = await this.getLayers(layersToGet ?? []);
+    const l = await this.getLayers(layersToGet ?? [], false);
     this.overlaysTree = [l.layers];
 
     flatMapDeep(this.overlaysTree.flat(), flattenOverlaysTree).forEach((x) => {
@@ -1103,11 +1107,11 @@ export class MapComponent
         const id = (x.layer as any).id;
         if (!isNil(shouldDisplayStatuses[id])) {
           (x.layer as any).shouldDisplay = shouldDisplayStatuses[id];
-          if (!shouldDisplayStatuses[id]) {
-            x.layer.remove();
-          } else {
+          if (shouldDisplayStatuses[id]) {
             this.map.addLayer(x.layer);
           }
+        } else {
+          this.map.addLayer(x.layer);
         }
       }
     });