+++ libs/shared/src/lib/components/widgets/summary-card/summary-card.component.ts
@@ -238,6 +238,7 @@ export class SummaryCardComponent
       );
       mapping = this.contextService.replaceContext(mapping);
       mapping = this.contextService.replaceFilter(mapping);
+      mapping = this.replaceWidgetVariables(mapping);
       this.contextService.removeEmptyPlaceholders(mapping);
       return mapping;
     } catch {
@@ -307,8 +308,12 @@ export class SummaryCardComponent
     ) {
       this.contextService.filter$
         .pipe(debounceTime(500), takeUntil(this.destroy$))
-        .subscribe(() => {
-          this.refresh();
+        .subscribe(({ previous, current }) => {
+          if (
+            this.contextService.shouldRefresh(this.widget, previous, current)
+          ) {
+            this.refresh();
+          }
         });
     }
   }
@@ -442,7 +447,7 @@ export class SummaryCardComponent
         Object.assign(
           this.graphqlVariables || {},
           this.refData.pageInfo?.pageSizeVar && {
-            [this.refData.pageInfo.pageSizeVar]: DEFAULT_PAGE_SIZE,
+            [this.refData.pageInfo.pageSizeVar]: this.pageInfo.pageSize,
           }
         )
       );
@@ -493,23 +498,27 @@ export class SummaryCardComponent
         })
         .then(this.updateRecordCards.bind(this));
     } else if (this.useReferenceData) {
-      const contextFilters = this.contextService.injectContext(
-        this.contextFilters
-      );
-      this.sortedCachedCards = cloneDeep(
-        this.cachedCards
-          .filter((x) => filterReferenceData(x.rawValue, contextFilters))
-          .filter((card: any) => {
-            return (
-              JSON.stringify(card.rawValue)
-                .replace(/("\w+":)/g, '')
-                .toLowerCase()
-                .indexOf(search.toLowerCase()) !== -1
-            );
-          })
-      );
-      this.cards = this.sortedCachedCards.slice(0, this.pageInfo.pageSize);
-      this.pageInfo.length = this.sortedCachedCards.length;
+      if (this.refData?.pageInfo?.strategy) {
+        this.refresh();
+      } else {
+        const contextFilters = this.contextService.injectContext(
+          this.contextFilters
+        );
+        this.sortedCachedCards = cloneDeep(
+          this.cachedCards
+            .filter((x) => filterReferenceData(x.rawValue, contextFilters))
+            .filter((card: any) => {
+              return (
+                JSON.stringify(card.rawValue)
+                  .replace(/("\w+":)/g, '')
+                  .toLowerCase()
+                  .indexOf(search.toLowerCase()) !== -1
+              );
+            })
+        );
+        this.cards = this.sortedCachedCards.slice(0, this.pageInfo.pageSize);
+        this.pageInfo.length = this.sortedCachedCards.length;
+      }
     }
   }
 
@@ -634,14 +643,6 @@ export class SummaryCardComponent
       })) as CardT[])
     );
 
-    const contextFilters = this.contextService.injectContext(
-      this.contextFilters
-    );
-
-    this.sortedCachedCards = cloneDeep(this.cachedCards).filter((x) =>
-      filterReferenceData(x.rawValue, contextFilters)
-    );
-
     const strategy = this.refData?.pageInfo?.strategy;
 
     const isPaginated = !!strategy && !!pageInfo;
@@ -654,7 +655,29 @@ export class SummaryCardComponent
       if (this.pageInfo.pageSize > items.length) {
         this.pageInfo.length = this.cards.length;
       }
+      this.sortedCachedCards = cloneDeep(this.cachedCards);
     } else {
+      // Client side filtering
+      const contextFilters = this.contextService.injectContext(
+        this.contextFilters
+      );
+
+      this.sortedCachedCards = cloneDeep(this.cachedCards).filter((x) =>
+        filterReferenceData(x.rawValue, contextFilters)
+      );
+
+      if (this.searchControl.value) {
+        this.sortedCachedCards = this.sortedCachedCards.filter((card: any) => {
+          return (
+            JSON.stringify(card.rawValue)
+              .replace(/("\w+":)/g, '')
+              .toLowerCase()
+              .indexOf((this.searchControl.value as string).toLowerCase()) !==
+            -1
+          );
+        });
+      }
+
       // If not, all the items are already loaded
       this.pageInfo.length = this.cachedCards.length;
     }
@@ -754,7 +777,7 @@ export class SummaryCardComponent
             this.dataQuery = this.apollo.watchQuery<any>({
               query: builtQuery,
               variables: {
-                first: DEFAULT_PAGE_SIZE,
+                first: this.pageInfo.pageSize,
                 filter: this.queryFilter,
                 sortField: this.sortOptions.field,
                 sortOrder: this.sortOptions.order,
@@ -871,7 +894,7 @@ export class SummaryCardComponent
     this.dataQuery = this.aggregationService.aggregationDataWatchQuery(
       card.resource as string,
       card.aggregation as string,
-      DEFAULT_PAGE_SIZE,
+      this.pageInfo.pageSize,
       0,
       this.contextService.injectContext(this.contextFilters),
       this.widget.settings.at
@@ -957,19 +980,20 @@ export class SummaryCardComponent
           }),
         })
         .then(this.updateRecordCards.bind(this));
-    } else if (this.useReferenceData) {
+    } else if (this.useReferenceData && this.refData) {
       const strategy = this.refData?.pageInfo?.strategy;
       const refData = this.refData;
+
+      // Build the variables for the query based on the pagination strategy
+      const variables: any = Object.assign(
+        {},
+        refData?.pageInfo?.pageSizeVar
+          ? { [refData.pageInfo.pageSizeVar]: this.pageInfo.pageSize }
+          : {}
+      );
+
       // If using pagination, fetch the next page
       if (strategy && refData?.pageInfo) {
-        // Build the variables for the query based on the pagination strategy
-        const variables: any = Object.assign(
-          {},
-          refData.pageInfo?.pageSizeVar
-            ? { [refData.pageInfo.pageSizeVar]: this.pageInfo.pageSize }
-            : {}
-        );
-
         // Set the pagination variable according to the strategy
         if (refData.pageInfo.strategy === 'offset') {
           variables[refData.pageInfo.offsetVar] = this.pageInfo.skip;
@@ -982,25 +1006,19 @@ export class SummaryCardComponent
         } else if (refData.pageInfo.strategy === 'page') {
           variables[refData.pageInfo.pageVar] = event.pageIndex + 1;
         }
-
-        // Only set loading state if using pagination, not infinite scroll
-        this.loading = !this.scrolling;
-        this.referenceDataService
-          .cacheItems(refData, {
-            ...variables,
-            ...(this.graphqlVariables ?? {}),
-          })
-          .then(({ items, pageInfo }) => {
-            this.updateReferenceDataCards(items, pageInfo);
-            this.loading = false;
-          });
-      } else {
-        this.cards = this.sortedCachedCards.slice(
-          this.pageInfo.skip,
-          this.pageInfo.skip + this.pageInfo.pageSize
-        );
-        this.pageInfo.length = this.sortedCachedCards.length;
       }
+
+      // Only set loading state if using pagination, not infinite scroll
+      this.loading = !this.scrolling;
+      this.referenceDataService
+        .cacheItems(refData, {
+          ...variables,
+          ...(this.graphqlVariables ?? {}),
+        })
+        .then(({ items, pageInfo }) => {
+          this.updateReferenceDataCards(items, pageInfo);
+          this.loading = false;
+        });
     }
   }
 
@@ -1008,40 +1026,20 @@ export class SummaryCardComponent
    * Refresh view
    */
   public refresh() {
-    // Only apply logic for reference data with no pagination
-    if (this.useReferenceData) {
-      if (!this.refData?.pageInfo?.strategy) {
-        const contextFilters = this.contextService.injectContext(
-          this.contextFilters
-        );
-        this.sortedCachedCards = cloneDeep(
-          this.cachedCards.filter((x) =>
-            filterReferenceData(x.rawValue, contextFilters)
-          )
-        );
-        if (this.searchControl.value) {
-          this.sortedCachedCards = this.sortedCachedCards.filter(
-            (card: any) => {
-              return (
-                JSON.stringify(card.rawValue)
-                  .replace(/("\w+":)/g, '')
-                  .toLowerCase()
-                  .indexOf(
-                    (this.searchControl.value as string).toLowerCase()
-                  ) !== -1
-              );
-            }
-          );
-        }
-      } else {
-        // Remove previous cached data
-        this.cards = [];
-        this.sortedCachedCards = [];
-        this.cachedCards = [];
-      }
+    this.cards = [];
+    this.sortedCachedCards = [];
+    this.cachedCards = [];
+
+    if (this.summaryCardGrid) {
+      this.summaryCardGrid.nativeElement.scroll({
+        top: 0,
+        left: 0,
+        behavior: 'smooth',
+      });
     }
+
     this.onPage({
-      pageSize: DEFAULT_PAGE_SIZE,
+      pageSize: this.pageInfo.pageSize,
       skip: 0,
       previousPageIndex: 0,
       pageIndex: 0,
@@ -1107,33 +1105,70 @@ export class SummaryCardComponent
           })
           .then(() => (this.loading = false));
       } else if (this.useReferenceData) {
-        this.loading = true;
-        this.pageInfo.pageIndex = 0;
-        this.pageInfo.skip = 0;
-        if (e) {
-          const field = `rawValue.${this.sortOptions.field as string}`;
-          if (this.sortOptions.order === 'asc') {
-            this.sortedCachedCards.sort((a, b) => {
-              const fieldA = String(get(a, field) || '');
-              const fieldB = String(get(b, field) || '');
-              return fieldA.localeCompare(fieldB);
-            });
-          } else {
-            this.sortedCachedCards.sort((a, b) => {
-              const fieldA = String(get(a, field) || '');
-              const fieldB = String(get(b, field) || '');
-              return fieldB.localeCompare(fieldA);
-            });
-          }
-          this.cards = this.sortedCachedCards.slice(0, this.pageInfo.pageSize);
+        if (this.refData?.pageInfo?.strategy) {
+          this.refresh();
         } else {
-          this.sortedCachedCards.sort(
-            (a, b) => (a.index as number) - (b.index as number)
-          );
-          this.cards = this.sortedCachedCards.slice(0, this.pageInfo.pageSize);
+          this.loading = true;
+          this.pageInfo.pageIndex = 0;
+          this.pageInfo.skip = 0;
+          if (e) {
+            const field = `rawValue.${this.sortOptions.field as string}`;
+            if (this.sortOptions.order === 'asc') {
+              this.sortedCachedCards.sort((a, b) => {
+                const fieldA = String(get(a, field) || '');
+                const fieldB = String(get(b, field) || '');
+                return fieldA.localeCompare(fieldB);
+              });
+            } else {
+              this.sortedCachedCards.sort((a, b) => {
+                const fieldA = String(get(a, field) || '');
+                const fieldB = String(get(b, field) || '');
+                return fieldB.localeCompare(fieldA);
+              });
+            }
+            this.cards = this.sortedCachedCards.slice(
+              0,
+              this.pageInfo.pageSize
+            );
+          } else {
+            this.sortedCachedCards.sort(
+              (a, b) => (a.index as number) - (b.index as number)
+            );
+            this.cards = this.sortedCachedCards.slice(
+              0,
+              this.pageInfo.pageSize
+            );
+          }
+          this.loading = false;
         }
-        this.loading = false;
       }
     }
+  }
+
+  /**
+   * Replace widget variables in mapping
+   *
+   * @param object mapping
+   * @returns updated mapping
+   */
+  private replaceWidgetVariables(object: any): any {
+    // Replace sort options
+    const sort = this.sortOptions;
+    if (sort && sort.field && sort.order) {
+      object = JSON.parse(
+        JSON.stringify(object)
+          .replace(/{{widget.sortField}}/g, sort.field)
+          .replace(/{{widget.sortOrder}}/g, sort.order)
+      );
+    }
+
+    // Replace search
+    const search = this.searchControl.value;
+    if (search) {
+      object = JSON.parse(
+        JSON.stringify(object).replace(/{{widget.search}}/g, search)
+      );
+    }
+    return object;
   }
 }